+++ src/GoogleMapsPlugin.php
@@ -17,6 +17,7 @@ use craft\base\Field;
 use craft\base\Model;
 use craft\base\Plugin;
 use craft\elements\Entry;
+use craft\events\ConfigEvent;
 use craft\events\ModelEvent;
 use craft\events\PluginEvent;
 use craft\events\RegisterComponentTypesEvent;
@@ -24,6 +25,7 @@ use craft\events\RegisterElementExportersEvent;
 use craft\helpers\UrlHelper;
 use craft\services\Fields;
 use craft\services\Plugins;
+use craft\services\ProjectConfig;
 use craft\services\Utilities;
 use doublesecretagency\googlemaps\exporters\AddressesCondensedExporter;
 use doublesecretagency\googlemaps\exporters\AddressesExpandedExporter;
@@ -96,6 +98,9 @@ class GoogleMapsPlugin extends Plugin
         // Register all events
         $this->_registerFieldType();
         $this->_registerExporters();
+
+        // Manage conversions of the Address field
+        $this->_manageFieldTypeConversions();
 
         // Redirect after plugin install
         $this->_postInstallRedirect();
@@ -177,6 +182,86 @@ class GoogleMapsPlugin extends Plugin
             static function(RegisterComponentTypesEvent $event) {
                 $event->types[] = TestGoogleApiKeysUtility::class;
                 $event->types[] = TestAddressLookupUtility::class;
+            }
+        );
+    }
+
+    // ========================================================================= //
+
+    /**
+     * Manage conversions of the Address field from the Mapbox plugin.
+     *
+     * @return void
+     */
+    private function _manageFieldTypeConversions(): void
+    {
+        // If Mapbox plugin is not installed, bail
+        if (!Craft::$app->getPlugins()->isPluginEnabled('mapbox')) {
+            return;
+        }
+
+        // When a single project config line gets updated
+        Event::on(
+            ProjectConfig::class,
+            ProjectConfig::EVENT_UPDATE_ITEM,
+            static function (ConfigEvent $event) {
+
+                // Get old and new types
+                $oldType = $event->oldValue['type'] ?? null;
+                $newType = $event->newValue['type'] ?? null;
+
+                // If old type wasn't a Mapbox Address field, bail
+                if (!($oldType === 'doublesecretagency\mapbox\fields\AddressField')) {
+                    return;
+                }
+
+                // If new type is not a Google Maps Address field, bail
+                if (!($newType === 'doublesecretagency\googlemaps\fields\AddressField')) {
+                    return;
+                }
+
+                // Get the field's UID
+                $uid = str_replace('fields.', '', $event->path);
+
+                // Get the actual field
+                $field = Craft::$app->getFields()->getFieldByUid($uid);
+
+                // If unable to get the field, bail
+                if (!$field) {
+                    return;
+                }
+
+                // List of columns to copy between tables
+                $columns = [
+                    'elementId', 'fieldId',
+                    'formatted', 'raw',
+                    'name', 'street1', 'street2',
+                    'city', 'state', 'zip',
+                    'county', 'country',
+                    'lat', 'lng', 'zoom',
+                    'dateCreated', 'dateUpdated', 'uid'
+                ];
+
+                // Merge and escape column names
+                $columns = '[['.implode(']],[[', $columns).']]';
+
+                // Copy field's rows from `mapbox_addresses` into `googlemaps_addresses`
+                $sql = "
+INSERT INTO [[googlemaps_addresses]] ({$columns})
+SELECT {$columns}
+FROM [[mapbox_addresses]]
+WHERE [[fieldId]] = :fieldId
+  AND NOT EXISTS (
+    SELECT 1
+    FROM [[googlemaps_addresses]]
+    WHERE [[googlemaps_addresses]].[[uid]] = [[mapbox_addresses]].[[uid]]
+)";
+
+                // Execute the SQL statement
+                \Yii::$app->db->createCommand($sql)
+                    ->bindValues([':fieldId' => $field->id])
+                    ->execute();
+
             }
         );
     }