+++ src/fields/AddressField.php
@@ -115,7 +115,7 @@ class AddressField extends Field implements PreviewableFieldInterface
      */
     public static function displayName(): string
     {
-        return Craft::t('google-maps', 'Address');
+        return Craft::t('google-maps', 'Address (Google Maps)');
     }
 
     /**