+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Added
 - Address fields can now be added to the search index. ([#21](https://github.com/doublesecretagency/craft-googlemaps/issues/21))
 - Optionally use minified JavaScript files for dynamic maps on the front-end. ([#98](https://github.com/doublesecretagency/craft-googlemaps/issues/98))
+- Added ability to [convert Address fields](https://plugins.doublesecretagency.com/google-maps/guides/converting-from-mapbox/) from the Mapbox plugin.
 
 ## 4.3.6 - 2023-09-18